pyproject.toml
[project]
name = "gp-search-grounding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.32.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
]
